blog/requirements.txt
Click==7.0
diff-match-patch==20181111
Django==2.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.2
django-simple-captcha==0.5.9
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.4
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.10
Pillow==5.3.0
pytz==2018.7
PyYAML==3.13
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
